requirements.txt
Flask
gunicorn
tensorflow==2.0.0
keras==2.3.1
opencv-python
pandas
pillow
numpy
